tablets: relocate tablets code to a separate package
Move code and tests to a separate package
