+++ tablets/tablets_test.go
@@ -1,7 +1,7 @@
 //go:build unit
 // +build unit
 
-package gocql
+package tablets
 
 import (
 	"testing"
